+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.86.1] - 2019-11-13
 ### Fixed
 - Bug of selected images of `ProductImages`, now the control of the active class is in the component instead of the `SwiperJS` lib.